flask_server/requirements.txt
Flask
pydub
SpeechRecognition
Flask-CORS
requests
mysql-connector-python==8.0.33
